refactor: inline adapter upsert helpers, improve command types, and extract test builders
Inline model upsert helpers directly into adapter create/update methods,
add type-safe required option support to command handlers, and extract
shared buildCustomType/buildSlice test builders into test/it.ts.
